add support for binary secret
This allows the engine to be used to watch and retrieve data
from binary secrets. The current secret implementation assumes
that the value is a string and hence does not work with binary
data.

Besides the fact that this is a template engine and binary data
make very little sense, this engine can be used as a generic
k8s watcher.

Related to CyCoreSystems/asterisk-config#8.

Signed-off-by: Vinicius Ruoso <vinicius@leads2b.com>