Use standard kubernetes client library
Migrated from unmaintained kubernetes client library to official one,
involving a major refactor.

**BREAKING CHANGES**: the invocation of this library has changed
substantially, and this is an incompatible change.

However, the template functions have stayed the same for this release.
We may yet change those in a later release to offer more flexibility and
be better aligned among themselves, but that will be later.

Signed-off-by: Seán C McCord <ulexus@gmail.com>